==> 01_Crawl_data.ipynb <==
from selenium import webdriver
import numpy as np
from time import sleep
import random
from selenium.common.exceptions import NoSuchElementException, ElementNotInteractableException
from selenium.webdriver.common.by import By
import pandas as pd
from time import sleep, time
import re, csv
from random import uniform, randint
from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.support.wait import WebDriverWait
from selenium.webdriver.common.action_chains import ActionChains
from selenium.webdriver.support import expected_conditions as EC
from selenium.common.exceptions import NoSuchElementException
from seleniumbase import SB
import undetected_chromedriver as uc


base_url = "https://batdongsan.com.vn/ban-nha-rieng-ha-noi/1pn/p{}?rs=2,3,4"
url_list = []
start_page = 21
end_page = 30   #20
for page in range(start_page, end_page +1):
    url = base_url.format(page)
    url_list.append(url)

url_list

# Function to get information

def get_info(link):
    driver.get(link)
    sleep(random.randint(5,10))
    while True:
#     Get links and title
        elems = driver.find_elements(By.CSS_SELECTOR, "div.js__card.js__card-full-web.pr-container.re__card-full.re__vip-normal [href]")
        title = [elem.text for elem in elems]
        links = [elem.get_attribute('href') for elem in elems]
#         link_demo = links
    #     Get price
        elems_price = driver.find_elements(By.CSS_SELECTOR, ".re__card-config-price.js__card-config-item")
        length = len(elems_price)
        price = [elem_price.text for elem_price in elems_price]

    #     Get location
        elems_loc = driver.find_elements(By.CSS_SELECTOR, ".re__card-location")
        # length_loc = len(elems_size)
        location = [elem_loc.text for elem_loc in elems_loc]

    #  Get size
        elems_size = driver.find_elements(By.CSS_SELECTOR, ".re__card-config-area.js__card-config-item")
        length_size = len(elems_size)
        size = [elem_size.text for elem_size in elems_size]

    #     Get percent price
    # <span class="re__card-config-price_per_m2 js__card-config-item">100 tr/m²</span>
        elems_per = driver.find_elements(By.CSS_SELECTOR, ".re__card-config-price_per_m2.js__card-config-item")
        length_per = len(elems_per)
        percent = [elem_per.text for elem_per in elems_per]

        df1 = pd.DataFrame(list(zip(title, price, links, location, size, percent)), columns = ['title', 'price','link_item', 'location', 'size', 'Price_per_m2'])
        df1['index_']= np.arange(1, len(df1) + 1)

    # Get bedroom
        bed_list, bed_idx = [], []
        for i in range(1, len(title)+1):
            try:
                bed = driver.find_element("xpath","/html/body/div[8]/div[2]/div[1]/div[6]/div[{}]/a/div[2]/div[1]/div[1]/div[1]/span[7]".format(i))
                bed_list.append(bed.text)
                bed_idx.append(i)
            except NoSuchElementException:
                pass

        df2 = pd.DataFrame(list(zip(bed_idx , bed_list)), columns = ['bed_idx', 'bedroom',])
        df3 = df1.merge(df2, how='left', left_on='index_', right_on='bed_idx')
    #     Get toilet
        toi_list, toi_idx = [], []
        for i in range(1, len(title)+1):
            try:
                toilet = driver.find_element("xpath","/html/body/div[8]/div[2]/div[1]/div[6]/div[{}]/a/div[2]/div[1]/div[1]/div[1]/span[9]".format(i))
                toi_list.append(toilet.text)
                toi_idx.append(i)
            except NoSuchElementException:
                pass
        df4 = pd.DataFrame(list(zip(toi_idx , toi_list)), columns = ['toi_idx', 'toi_list'])
        df5 = df3.merge(df4, how='left', left_on='index_', right_on='toi_idx')
        driver.close()
        return df5
    

all_data = []
for link in url_list:
    options = uc.ChromeOptions() 
    options.headless = False  
    driver = uc.Chrome(use_subprocess=True, options=options) 
    driver.get(link) 
    driver.maximize_window() 
    df = get_info(link)
    all_data.append(df)
    final_df = pd.concat(all_data, ignore_index=True )
    sleep(3)
    
# driver.close()

final_df

meo = []
meo = final_df['link_item']


#  example

meo1 = meo[1:5]
meo1

def getDetailItems(link):
    # driver = webdriver.Chrome(executable_path='/path/to/chromedriver')  # Replace with the path to your ChromeDriver executable
    driver.get(link)
    data = {
        "link_item": link,
        "mat_tien": None,
        "floor": None,
        "duong_vao": None,
        "phap_ly" :None,
        "noi_that" : None,
        "Huong" :None,
        "Huong_ban_cong": None,
    }

    # Find all elements with class "re__pr-specs-content-item"
    elems = driver.find_elements(By.CSS_SELECTOR, ".re__pr-specs-content-item")

    for elem in elems:
        title = elem.find_element(By.CLASS_NAME, "re__pr-specs-content-item-title").text
        value = elem.find_element(By.CLASS_NAME, "re__pr-specs-content-item-value").text

        if title == "Mặt tiền":
            data["mat_tien"] = value
        elif title == "Số tầng":
            data["floor"] = value
        elif title == "Đường vào":
            data["duong_vao"] = value
        elif title == "Pháp lý":
            data["phap_ly"] = value
        elif title == "Nội thất":
            data["noi_that"] = value
        elif title == "Hướng nhà":
            data["Huong"] = value
        elif title == "Hướng ban công":
            data["Huong_ban_cong"] = value

    df0 = pd.DataFrame([data])
    driver.close()
    return df0

# meo1 = ["url1", "url2"]  # Replace with your list of URLs
df_list = []

for link in meo:
    options = uc.ChromeOptions() 
    options.headless = False  # Set headless to False to run in non-headless mode

    driver = uc.Chrome(use_subprocess=True, options=options) 
    driver.get(link) 
    driver.maximize_window() 

    df = getDetailItems(link)
    df_list.append(df)
    sleep(3)

final_meo = pd.concat(df_list, ignore_index=True)

# Save the final DataFrame to a CSV file
# final_meo.to_csv('output.csv', index=False)

final_meo

concatenated_df = pd.concat([final_df, final_meo], axis=1)

# Make sure the indices align or reset them if needed
final_df.reset_index(drop=True, inplace=True)
final_meo.reset_index(drop=True, inplace=True)

# Then, concatenate horizontally
concatenated_df = pd.concat([final_df, final_meo], axis=1)

concatenated_df

concatenated_df.to_csv('house4.csv')
